+++ app/src/main/java/com/tang/prm/di/DatabaseModule.kt
@@ -44,7 +44,8 @@ object DatabaseModule {
                 MIGRATION_26_27,
                 MIGRATION_27_28,
                 MIGRATION_28_29,
-                MIGRATION_29_30
+                MIGRATION_29_30,
+                MIGRATION_30_31
             )
             .build()
     }